requirements.txt
bcrypt==3.2.0
black==21.9b0
cffi==1.14.6
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
pycparser==2.20
python-slugify==5.0.2
regex==2021.8.28
six==1.16.0
SQLAlchemy==1.4.23
text-unidecode==1.3
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
WTForms==2.3.3
